% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspartyPPtreereg.R
\name{as_party}
\alias{as_party}
\title{Party Object for projection pursuit regression tree}
\usage{
as_party(PPTreeregOBJ, Rule=1, data=TRUE,...)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{Rule}{split rule
1: mean of two group means
2: weighted mean of two group means - weight with group size
3: weighted mean of two group means - weight with group sd
4: weighted mean of two group means - weight with group se
5: mean of two group medians
6: weighted mean of two group medians - weight with group size
7: weighted mean of two group median - weight with group IQR
8: weighted mean of two group median - weight with group IQR
                                       and group size
9: cutoff that minimize error rates in each node}

\item{data}{a (potentially empty) data.frame.}

\item{...}{arguments to be passed to methods}
}
\description{
Change to party object, only conducted when final.rule = 1
}
