% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulated.R
\docType{data}
\name{dataXY}
\alias{dataXY}
\title{Simulated data}
\format{
A data frame with 100 rows and 4 variables.
}
\usage{
data(dataXY)
}
\description{
The \code{dataXY} dataset is simulated data for running Projection Pursuit regression tree model.
}
\details{
It contains 100 rows and 4 variables.
}
\references{
\doi{https://doi.org/10.3390/app11219885}
}
\keyword{dataXY}
