% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_decision.R
\name{decisionplot}
\alias{decisionplot}
\title{decisionplot}
\usage{
decisionplot(
  PPTreeregOBJ,
  testObs,
  final.rule = 5,
  method = "simple",
  varImp = "shapImp"
)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{testObs}{testObs}

\item{final.rule}{final.rule}

\item{method}{method}

\item{varImp}{shapImp or treeimp}
}
\description{
decisionplot
}
