% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_insurance.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{insurance data}
\format{
a data frame with 1338 rows and 7 columns.
}
\source{
The \code{insurance.csv} dataset was downloaded from the Kaggle site.
The dataset was obtained from
\url{https://www.kaggle.com/mirichoi0218/insurance} on May 11, 2021.
}
\usage{
data(insurance)
}
\description{
The \code{insurance} dataset is a preprocessed \code{insurance.csv} dataset which comes as
a part of "Medical Cost Personal Datasets" at Kaggle.
}
\details{
It contains 1338 rows and 7 variables. These are:

\itemize{
\item age - age of primary beneficiary
\item sex - insurance contractor gender, female, male
\item bmi - Body mass index, providing an understanding of body, weights that are relatively high or low relative to height,
objective index of body weight (kg / m ^ 2) using the ratio of height to weight, ideally 18.5 to 24.9
\item children - Number of children covered by health insurance / Number of dependents
\item smoker - Smoking
\item region - the beneficiary's residential area in the US, northeast, southeast, southwest, northwest. charges: Individual medical costs billed by health insurance
}


Source: \url{https://www.kaggle.com/mirichoi0218/insurance}
}
\keyword{insurance}
