% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{observation_impute_cpp_simple}
\alias{observation_impute_cpp_simple}
\title{Get imputed data}
\usage{
observation_impute_cpp_simple(xbar, index_simple, xtest, S)
}
\arguments{
\item{xbar}{mean of each leaf.}

\item{index_simple}{Positive integer.}

\item{xtest}{Numeric matrix. Represents a single test observation.}

\item{S}{Integer matrix of dimension \code{n_combinations x m}, where \code{n_combinations} equals
the total number of sampled/non-sampled feature combinations and \code{m} equals
the total number of unique features. Note that \code{m = ncol(xtrain)}. See details
for more information.}
}
\value{
Numeric matrix
}
\description{
Get imputed data
}
\author{
Nikolai Sellereite
}
\keyword{internal}
