% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPPtreereg.R
\name{plot.PPTreereg}
\alias{plot.PPTreereg}
\title{PPTreereg plot}
\usage{
\method{plot}{PPTreereg}(x, font.size = 17, width.size = 1, ...)
}
\arguments{
\item{x}{PPTreereg object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{...}{arguments to be passed to methods}
}
\description{
Projection pursuit regression tree plot
}
\details{
plot projection pursuit regression tree structure
}
\keyword{tree}
