% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPPggparty.R
\name{pp_ggparty}
\alias{pp_ggparty}
\title{Plot using ggparty}
\usage{
pp_ggparty(PPTreeregOBJ,ind_variable,...)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{ind_variable}{independent varible to show}

\item{...}{arguments to be passed to methods}
}
\description{
Put Indipendent Variable
}
\keyword{tree}
