% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPimportance.R
\name{PPimportance}
\alias{PPimportance}
\title{PPimportance PPTreereg}
\usage{
PPimportance(PPTreeregOBJ,...)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{...}{arguments to be passed to methods}
}
\value{
imp_node_split predicted values

imp_node_final MSE of the predicted values

imp_var MSE of the predicted values
}
\description{
Calculate importance
}
\keyword{tree}
