% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPregCoeff.R
\name{PPregcoeff}
\alias{PPregcoeff}
\title{predict PPtree}
\usage{
PPregcoeff(PPTreeregOBJ,c,...)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{c}{...}

\item{...}{arguments to be passed to methods}
}
\value{
coef_result coefficients of nodes
}
\description{
Predict class for the test set and calculate prediction error
After finding tree structure, predict class for the test set and calculate
prediction error.
}
\examples{
data(mtcars)
n <- nrow(mtcars)
tot <- c(1:n)
n.train <- round(n*0.8)
train <- sample(tot,n.train)
test <- tot[-train]
Tree.result <- PPTreereg(mpg~.,mtcars[train,],
                           final.rule=1,DEPTH=2)
PPregcoeff(Tree.result,c(1:3))
}
\keyword{tree}
