% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPregNodeViz.R
\name{PPregNodeViz}
\alias{PPregNodeViz}
\title{Visualization of each node in PPtree}
\usage{
PPregNodeViz(PPTreeregOBJ,node.id,Rule=5)
}
\arguments{
\item{PPTreeregOBJ}{PPregtree object}

\item{node.id}{node ID}

\item{Rule}{cutoff rule}
}
\description{
Explore PPtree with different Rules in each split.
}
\keyword{tree}
