% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPshapdependence.R
\name{PPshapdependence}
\alias{PPshapdependence}
\title{Title}
\usage{
PPshapdependence(data_long, x, y = NULL, color_feature = NULL, smooth = TRUE)
}
\arguments{
\item{data_long}{ppshapr_prep object}

\item{x}{variable1}

\item{y}{variable2}

\item{color_feature}{display color}

\item{smooth}{geom_smooth}
}
\description{
Title
}
