% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapPP.R
\name{ppshapr.empirical}
\alias{ppshapr.empirical}
\title{Calculate \code{PPKernelSHAP} values with empirical methods}
\usage{
ppshapr.empirical(PPTreeregOBJ, testObs, final.rule, final.leaf = NULL)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg class object - a model to be explained}

\item{testObs}{test data observation}

\item{final.rule}{final rule to assign numerical values in the final nodes.
1: mean value in the final nodes
2: median value in the final nodes
3: using optimal projection
4: using all independent variables
5: using several significant independent variables}

\item{final.leaf}{location of final leaf}
}
\description{
Calculate \code{PPKernelSHAP} values with empirical methods
}
