% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppshap_calculate.R
\name{ppshapr.simple}
\alias{ppshapr.simple}
\title{Calculate shap values with simple methods}
\usage{
ppshapr.simple(PPTreeregOBJ, testObs, final.rule)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{testObs}{testObs}

\item{final.rule}{rule to calculate the final node value}
}
\description{
Calculate shap values with simple methods
}
