% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPshapsummary.R
\name{PPshapsummary}
\alias{PPshapsummary}
\title{Title}
\usage{
PPshapsummary(data_long, ...)
}
\arguments{
\item{data_long}{ppshapr_prep object}

\item{...}{arguments to be passed to methods}
}
\description{
Title
}
