% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainPPModel.R
\docType{data}
\name{PPTreereg.M1}
\alias{PPTreereg.M1}
\title{model method in caret}
\format{
An object of class \code{list} of length 11.
}
\usage{
PPTreereg.M1
}
\description{
PPTreereg Models in train
}
\details{
train in caret package
}
\keyword{datasets}
