% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainPPModel.R
\docType{data}
\name{PPTreereg.M5}
\alias{PPTreereg.M5}
\title{\code{PPTreereg} Models with \code{final.rule=5}}
\format{
An object of class \code{list} of length 11.
}
\usage{
PPTreereg.M5
}
\description{
Load \code{PPTreereg} Models to train
}
\details{
Load \code{PPTreereg} models to make explanation with caret package.
}
\examples{
\donttest{
data(dataXY)
PP_model <- caret::train(Y ~., data = dataXY, method = PPTreereg.M5, DEPTH=2, PPmethod="LDA")
}
}
\keyword{datasets}
