% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPregpredict.R
\name{predict.PPTreereg}
\alias{predict.PPTreereg}
\alias{predict}
\title{predict PPregReg}
\usage{
\method{predict}{PPTreereg}(
  object,
  newdata = NULL,
  Rule = 1,
  final.rule = 1,
  classinfo = FALSE,
  ...
)
}
\arguments{
\item{object}{PPtreereg object}

\item{newdata}{the test dataset}

\item{Rule}{split rule
1: mean of two group means
2: weighted mean of two group means - weight with group size
3: weighted mean of two group means - weight with group sd
4: weighted mean of two group means - weight with group se
5: mean of two group medians
6: weighted mean of two group medians - weight with group size
7: weighted mean of two group median - weight with group IQR
8: weighted mean of two group median - weight with group IQR
                                       and group size
9: cutoff that minimize error rates in each node}

\item{final.rule}{rule to calculate the final node value}

\item{classinfo}{return final node information (default is FALSE)}

\item{...}{arguments to be passed to methods}
}
\description{
predict PPtree
}
\details{
Predict class for the test set and calculate prediction error
After finding tree structure, predict class for the test set and calculate
prediction error.
}
\keyword{tree}
