% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPPtreereg.R
\name{print.PPTreereg}
\alias{print.PPTreereg}
\title{Print PPTreereg result}
\usage{
\method{print}{PPTreereg}(
  x,
  tree.print = TRUE,
  coef.print = FALSE,
  cutoff.print = FALSE,
  verbose = TRUE,
  final.rule = 1,
  ...
)
}
\arguments{
\item{x}{PPTreereg object}

\item{tree.print}{print the tree structure when TRUE}

\item{coef.print}{print the projection coefficient in each node when TRUE}

\item{cutoff.print}{print the cutoff values in each node when TRUE}

\item{verbose}{print if TRUE, no output if FALSE}

\item{final.rule}{rule to calculate the final node value}

\item{...}{arguments to be passed to methods}
}
\value{
tree print
}
\description{
Print PP.Tree.reg result
}
\details{
Print the projection pursuit regression tree result
}
\keyword{tree}
