% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubmodularPick.R
\name{SubPick_PPTreereg}
\alias{SubPick_PPTreereg}
\title{projection pursuit \code{submodular} pick algorithm \code{PP SP-LIME}}
\usage{
SubPick_PPTreereg(PPTreeregOBJ, explainer, chooseClass, obsnum =1,... )
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg class object - a model to be explained}

\item{explainer}{explanation object that generated by \code{lime} package.}

\item{chooseClass}{final node name, such as \code{"class1"}}

\item{obsnum}{The number of budgets (instance to be selected). Default value is 1.}

\item{...}{arguments to be passed to methods}
}
\value{
Observation names and their original values as data
}
\description{
Pick several data containing various information for each final node for \code{PPTreereg}
}
\details{
\code{submodular} Pick (\code{SP-LIME}) was developed (Ribeiro et al., 2016) to selects
representative data with important information to determine the
reliability of model based on the \code{LIME} algorithm.
In order to extract data for each final node in the \code{PPTreereg} model,
 \code{PP SP-LIME} was proposed based on \code{SP-LIME}.
}
\examples{
\donttest{
data(dataXY)
Model <- PPTreereg(Y~., data = dataXY, DEPTH = 2)
PP_model <- caret::train(Y~., data = dataXY, method = PPTreereg.M3, DEPTH=2, PPmethod="LDA")
lime_explainer <- lime::lime(dataXY[,-1],PP_model)
SubPick_PPTreereg(Model, lime_explainer, "class1", obsnum = 3)
}
}
\references{
Ribeiro, Marco Tulio, Sameer Singh, and Carlos Guestrin.
"" Why should i trust you?" Explaining the predictions of any classifier." Proceedings of the 22nd ACM SIGKDD international conference on knowledge discovery and data mining. 2016.
\doi{10.1145/2939672.2939778}

\url{https://github.com/marcotcr/lime/blob/master/lime/submodular_pick.py}
}
\keyword{submodular}
