% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.PPTreereg.R
\name{summary.PPTreereg}
\alias{summary.PPTreereg}
\title{Summary \code{PPTreereg} result}
\usage{
\method{summary}{PPTreereg}(object, c = NA, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from \code{PPTreereg}}

\item{c}{choose node id to summary. Default value is FALSE.}

\item{...}{arguments to be passed to methods}
}
\description{
summary \code{PPTreereg} result
}
\details{
summary the projection pursuit regression tree result
}
