% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppshap_calculate.R
\name{weight_matrix}
\alias{weight_matrix}
\title{weight_matrix}
\usage{
weight_matrix(X, normalize_W_weights = TRUE)
}
\arguments{
\item{X}{X}

\item{normalize_W_weights}{TRUE}
}
\description{
weight_matrix
}
