% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_PP.R
\name{explain_PP}
\alias{explain_PP}
\title{Make explainer for PPTreeregObj}
\usage{
explain_PP(PPTreeregOBJ, data, y, final.rule,...)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{data}{data.frame or matrix - data that was used for fitting. If not provided then will be extracted from the model. Data should be passed without target column (this shall be provided as the y argument).}

\item{y}{numeric vector with outputs / scores. If provided then it shall have the same size as data}

\item{final.rule}{rule to calculate the final node value}

\item{...}{arguments to be passed to methods}
}
\description{
explainer make for PPTreereg
}
\examples{
\dontrun{library(DALEX)
library(PPtreeRegViz)
data(mtcars)
PP_Tree <- PPtreeregViz::PPTreereg(mpg~.,data=mtcars)
explainer_PPTreereg <- explain_PP(model = PP_Tree,
data = mtcars[, -1], y = mtcars[,1], final.rule = 4)
# model performance plot
DALEX::model_performance(explainer_PPTreereg) \%>\% plot()
#Break Down Profile
DALEX::predict_parts(explainer_PPTreereg, new_observation =  mtcars[1,]) \%>\% plot()
}
}
