% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppshap_calculate.R
\name{feature_exact}
\alias{feature_exact}
\title{feature_exact}
\usage{
feature_exact(m, weight_zero_m = 10^6)
}
\arguments{
\item{m}{m}

\item{weight_zero_m}{weight_zero_m}
}
\description{
feature_exact
}
