% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapPP.R
\name{feature_exact}
\alias{feature_exact}
\title{feature_exact}
\usage{
feature_exact(m, weight_zero_m = 10^6)
}
\arguments{
\item{m}{m}

\item{weight_zero_m}{weight_zero_m}
}
\description{
feature_exact
}
\references{
The \code{shapr} package developed by
Nikolai Sellereite, Martin Jullum, Annabelle Redelmeier, Norsk Regnesentral.
\doi{10.1016/j.artint.2021.103502} and modified some codes at
\url{https://github.com/NorskRegnesentral/shapr}
}
\author{
Nikolai Sellereite
}
