% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPPimportance.R
\name{plot.PPimportance}
\alias{plot.PPimportance}
\title{Plot Importance of trained Tree}
\usage{
\method{plot}{PPimportance}(x, marginal = FALSE, num_var = 5, ...)
}
\arguments{
\item{x}{PPTreereg Importance object}

\item{marginal}{show each node. default if FALSE}

\item{num_var}{number of variables to show}

\item{...}{arguments to be passed to methods}
}
\description{
Plot Importance of trained Tree
}
\keyword{tree}
