% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPPtreereg.R
\name{plot.PPTreereg}
\alias{plot.PPTreereg}
\alias{plot}
\title{PPTreereg plot}
\usage{
\method{plot}{PPTreereg}(x, font.size = 17, width.size = 1, ...)
}
\arguments{
\item{x}{PPTreereg class object}

\item{font.size}{font size of plot}

\item{width.size}{size of eclipse in each node.}

\item{...}{arguments to be passed to methods}
}
\description{
projection pursuit regression tree plot
}
\details{
Draw projection pursuit regression tree with tree structure. It is modified from a function in \code{party} library.
}
\examples{
data(dataXY)
Model <- PPTreereg(Y~., data = dataXY, DEPTH = 2)
plot(Model)

}
\keyword{tree}
