% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPregCoeff.R
\name{PPregcoeff}
\alias{PPregcoeff}
\title{predict PPtree}
\usage{
PPregcoeff(PPTreeregOBJ,c,...)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{c}{...}

\item{...}{arguments to be passed to methods}
}
\value{
coef_result coefficients of nodes
}
\description{
Predict class for the test set and calculate prediction error
After finding tree structure, predict class for the test set and calculate
prediction error.
}
\keyword{tree}
