% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPregVarViz.R
\name{PPregVarViz}
\alias{PPregVarViz}
\title{Visualization of each node in PPtree}
\usage{
PPregVarViz(PPTreeregOBJ,var.id,indiv=FALSE,
                   DEPTH=NULL,smoothMethod="auto")
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{var.id}{X variable name}

\item{indiv}{TRUE: individual group plot, FLASE: combined one plot}

\item{DEPTH}{depth for exploration}

\item{smoothMethod}{method in geom_smooth function}
}
\description{
Explore PPtree with different Rules in each split.
}
\keyword{tree}
