% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppshap_calculate.R
\name{ppshapr.empirical}
\alias{ppshapr.empirical}
\title{Calculate shap values with empirical methods}
\usage{
ppshapr.empirical(PPTreeregOBJ, testObs, final.rule)
}
\arguments{
\item{PPTreeregOBJ}{PPTreeregOBJ}

\item{testObs}{testObs}

\item{final.rule}{rule to calculate the final node value}
}
\description{
Calculate shap values with empirical methods
}
