% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppshapr_prep.R
\name{ppshapr_prep}
\alias{ppshapr_prep}
\title{Title}
\usage{
ppshapr_prep(PPTreeshapObj = NULL, final.rule = 5, method = "empirical")
}
\arguments{
\item{PPTreeshapObj}{PPTreereg object}

\item{final.rule}{final.rule}

\item{method}{simple or empirical}
}
\description{
Title
}
