% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapPPsummary.R
\name{PPshapsummary}
\alias{PPshapsummary}
\title{Summary plot}
\usage{
PPshapsummary(data_long,...)
}
\arguments{
\item{data_long}{\code{ppshapr_prep} class object.}

\item{...}{arguments to be passed to methods}
}
\value{
An object of the class \code{ggplot}
}
\description{
Summary plot using \code{PPKernelSHAP}
}
\details{
A summary plot is used to see the aspects of important variables for each final node.
The summary plot summarizes information about the independent variables that contributed the most to the model's prediction
in the training data in the form of a density plot.
}
\examples{
data(dataXY)
testX <- dataXY[1,-1]
Model <- PPTreereg(Y~., data = dataXY, DEPTH = 2)
shap_long <- ppshapr_prep(Model, final.rule =5, method="simple")
PPshapsummary(shap_long)

}
