% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapPP.R
\name{shapley_weights}
\alias{shapley_weights}
\title{shapley_weights}
\usage{
shapley_weights(m, N, n_components, weight_zero_m = 10^6)
}
\arguments{
\item{m}{m}

\item{N}{N}

\item{n_components}{n_components}

\item{weight_zero_m}{weight_zero_m}
}
\description{
shapley_weights
}
\references{
The \code{shapr} package developed by
Nikolai Sellereite, Martin Jullum, Annabelle Redelmeier, Norsk Regnesentral.
\doi{10.1016/j.artint.2021.103502} and modified some codes at
\url{https://github.com/NorskRegnesentral/shapr}
}
\author{
Nikolai Sellereite
}
