% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPSubmodularPick.R
\name{SubPick_PPTreereg}
\alias{SubPick_PPTreereg}
\title{Print sublime observations from choosed Class}
\usage{
SubPick_PPTreereg(PPTreeregOBJ, explainer, chooseClass, obsnum = 1, ...)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{explainer}{explainer}

\item{chooseClass}{ex class1}

\item{obsnum}{default 1}

\item{...}{arguments to be passed to methods}
}
\description{
Pick obs according to class range
}
\details{
Pick obs according to class range
}
\examples{
\dontrun{modelPP <- caret::train(mpg~.,data =mtcars,
method = PPTreereg.M1 ,
DEPTH=2,PPmethod="LDA")
explainer_PP <- lime::lime(mtcars[,-1], modelPP)
SubPick_PPTreereg(PP_Tree, explainer_PP, "class1", 3)
}
}
\keyword{submodular}
