% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPPtreereg.R
\name{summary.PPTreereg}
\alias{summary.PPTreereg}
\title{Summary PPTreereg result}
\usage{
\method{summary}{PPTreereg}(object, c = NA, ...)
}
\arguments{
\item{object}{PPTreereg object}

\item{c}{number of node}

\item{...}{arguments to be passed to methods}
}
\description{
summary PPTreereg result
}
\details{
summary the projection pursuit regression tree result
}
