% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_decision.R
\name{waterfallplot}
\alias{waterfallplot}
\title{waterfallplot}
\usage{
waterfallplot(PPTreeregOBJ, testObs, final.rule = 5, method = "simple")
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{testObs}{testObs}

\item{final.rule}{final.rule}

\item{method}{simple or empirical}
}
\description{
waterfallplot
}
