% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_decision.R
\name{waterfallplot}
\alias{waterfallplot}
\title{waterfallplot}
\usage{
waterfallplot(
  PPTreeregOBJ,
  testObs,
  final.rule = 5,
  method = "simple",
  final.leaf = NULL
)
}
\arguments{
\item{PPTreeregOBJ}{PPTreereg object}

\item{testObs}{testObs}

\item{final.rule}{final.rule}

\item{method}{simple or empirical}

\item{final.leaf}{final leaf}
}
\description{
waterfallplot
}
