% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapPP.R
\name{weight_matrix}
\alias{weight_matrix}
\title{weight_matrix}
\usage{
weight_matrix(X, normalize_W_weights = TRUE)
}
\arguments{
\item{X}{X}

\item{normalize_W_weights}{default is TRUE}
}
\description{
weight_matrix
}
\references{
The \code{shapr} package developed by
Nikolai Sellereite, Martin Jullum, Annabelle Redelmeier, Norsk Regnesentral.
\doi{10.1016/j.artint.2021.103502} and modified some codes at
\url{https://github.com/NorskRegnesentral/shapr}
}
\author{
Nikolai Sellereite
}
