% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapPP.R
\name{weight_matrix}
\alias{weight_matrix}
\title{weight_matrix}
\usage{
weight_matrix(X, normalize_W_weights = TRUE)
}
\arguments{
\item{X}{X}

\item{normalize_W_weights}{default is TRUE}
}
\value{
Numeric matrix
}
\description{
The original source for much of this came from 'shapr' package
code in github.com/NorskRegnesentral/shapr/blob/master/R/shapley.R
Below is the original license statement for 'shapr' package.
}
\details{
MIT License
Copyright (c) 2019 Norsk Regnesentral
Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
}
\references{
The \code{shapr} package developed by
Nikolai Sellereite, Martin Jullum, Annabelle Redelmeier, Norsk Regnesentral.
\doi{10.1016/j.artint.2021.103502} and modified some codes at
\url{https://github.com/NorskRegnesentral/shapr}
}
\author{
Nikolai Sellereite
}
